+++ EpicorRESTGenerator.Shared/ClientBase.cs
