+++ EpicorRESTGenerator.Shared/Models/Title.cs
@@ -1,4 +1,4 @@
-﻿namespace EpicorSwaggerRESTGenerator.Models
+﻿namespace EpicorRESTGenerator.Shared.Models
 {
     [System.Xml.Serialization.XmlType(AnonymousType = true, Namespace = "http://www.w3.org/2005/Atom")]
     [System.Xml.Serialization.XmlRoot(Namespace = "http://www.w3.org/2005/Atom", IsNullable = false)]
