+++ EpicorRESTGenerator.Shared/app.config
