+++ EpicorRESTGenerator.Shared/packages.config
