+++ EpicorRESTGenerator/App.config
@@ -2,7 +2,7 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="EpicorRESTGenerator.WPFGUI.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false"/>
+            <section name="EpicorRESTGenerator.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false"/>
         </sectionGroup>
     </configSections>
     <startup> 
@@ -17,7 +17,7 @@
     </assemblyBinding>
   </runtime>
   <userSettings>
-    <EpicorRESTGenerator.WPFGUI.Properties.Settings>
+    <EpicorRESTGenerator.Properties.Settings>
       <setting name="EpicorUrl" serializeAs="String">
         <value>https://yourepicorwebsite.com/instancename</value>
       </setting>
@@ -36,6 +36,6 @@
       <setting name="UseCredentials" serializeAs="String">
         <value>True</value>
       </setting>
-    </EpicorRESTGenerator.WPFGUI.Properties.Settings>
+    </EpicorRESTGenerator.Properties.Settings>
   </userSettings>
 </configuration>
