+++ EpicorRESTGenerator/App.xaml
@@ -1,8 +1,8 @@
-﻿<Application x:Class="EpicorSwaggerRESTGenerator.WPFGUI.App"
+﻿<Application x:Class="EpicorRESTGenerator.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:properties="clr-namespace:EpicorRESTGenerator.WPFGUI.Properties"
-             xmlns:local="clr-namespace:EpicorSwaggerRESTGenerator.WPFGUI"
+             xmlns:properties="clr-namespace:EpicorRESTGenerator.Properties"
+             xmlns:local="clr-namespace:EpicorRESTGenerator"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
         <properties:Settings x:Key="Settings" />
