+++ EpicorRESTGenerator/App.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows;
 
-namespace EpicorSwaggerRESTGenerator
+namespace EpicorRESTGenerator
 {
     /// <summary>
     /// Interaction logic for App.xaml
