+++ EpicorRESTGenerator/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="EpicorSwaggerRESTGenerator.WPFGUI.MainWindow"
+﻿<Window x:Class="EpicorRESTGenerator.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:EpicorSwaggerRESTGenerator.WPFGUI"
+        xmlns:local="clr-namespace:EpicorRESTGenerator"
         mc:Ignorable="d"
         Title="Epicor Client Generator" 
         Height="800"
