+++ EpicorRESTGenerator/Properties/Resources.resx
